# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Controle de Automóveis",
    "version": "1.0.0",
    "description": "Documentação da API para cadastro e controle de utilização de veículos por motoristas."
  },
  "paths": {
    "/vehicles": {
      "get": {
        "summary": "Listar veículos",
        "parameters": [
          { "name": "color", "in": "query", "schema": { "type": "string" } },
          { "name": "brand", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Lista de veículos" } }
      },
      "post": {
        "summary": "Cadastrar veículo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plate": { "type": "string" },
                  "color": { "type": "string" },
                  "brand": { "type": "string" }
                },
                "required": ["plate", "color", "brand"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Veículo cadastrado" } }
      }
    },
    "/vehicles/{id}": {
      "get": {
        "summary": "Buscar veículo por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Veículo encontrado" } }
      },
      "put": {
        "summary": "Atualizar veículo",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": { "200": { "description": "Veículo atualizado" } }
      },
      "delete": {
        "summary": "Excluir veículo",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Veículo excluído" } }
      }
    },
    "/drivers": {
      "get": {
        "summary": "Listar motoristas",
        "parameters": [
          { "name": "name", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Lista de motoristas" } }
      },
      "post": {
        "summary": "Cadastrar motorista",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Motorista cadastrado" } }
      }
    },
    "/drivers/{id}": {
      "get": {
        "summary": "Buscar motorista por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Motorista encontrado" } }
      },
      "put": {
        "summary": "Atualizar motorista",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": { "200": { "description": "Motorista atualizado" } }
      },
      "delete": {
        "summary": "Excluir motorista",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Motorista excluído" } }
      }
    },
    "/usages": {
      "get": {
        "summary": "Listar utilizações",
        "responses": { "200": { "description": "Lista de utilizações de veículos" } }
      },
      "post": {
        "summary": "Registrar uso de veículo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vehicleId": { "type": "integer" },
                  "driverId": { "type": "integer" },
                  "reason": { "type": "string" }
                },
                "required": ["vehicleId", "driverId", "reason"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Uso registrado com sucesso" },
          "400": { "description": "Veículo ou motorista indisponível" }
        }
      }
    },
    "/usages/{id}/finalizar": {
      "patch": {
        "summary": "Finalizar uso de veículo",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Uso finalizado" },
          "400": { "description": "Erro ao finalizar uso" }
        }
      }
    }
  }
}